+++ tybotRoute/tiledeskChatbotPlugs/DirectivesChatbotPlug.js
@@ -25,6 +25,7 @@ const { DirCondition } = require('./directives/DirCondition');
 const { DirJSONCondition } = require('./directives/DirJSONCondition');
 const { DirAssign } = require('./directives/DirAssign');
 const { DirSetAttribute } = require('./directives/DirSetAttribute');
+const { DirSetAttributeV2 } = require('./directives/DirSetAttributeV2');
 const { DirWebRequest } = require('./directives/DirWebRequest');
 const { DirWebRequestV2 } = require('./directives/DirWebRequestV2');
 const { DirCode } = require('./directives/DirCode');
@@ -39,6 +40,7 @@ const { DirRandomReply } = require('./directives/DirRandomReply');
 const { DirGptTask } = require('./directives/DirGptTask');
 const { DirForm } = require('./directives/DirForm');
 const { DirCaptureUserReply } = require('./directives/DirCaptureUserReply');
+const { DirMake } = require('./directives/DirMake');
 
 class DirectivesChatbotPlug {
 
@@ -370,6 +372,13 @@ class DirectivesChatbotPlug {
         this.process(next_dir);
       });
     }
+    else if (directive_name === Directives.SET_ATTRIBUTE_V2) {
+      // console.log("...DirSetAttribute");
+      new DirSetAttributeV2(context).execute(directive, async () => {
+        let next_dir = await this.nextDirective(this.directives);
+        this.process(next_dir);
+      });
+    }
     // else if (directive_name === Directives.WHEN_OPEN) {
     //   // DEPRECATED
     //   const whenOpenDir = new DirWhenOpen(
@@ -611,6 +620,12 @@ class DirectivesChatbotPlug {
     }
     else if (directive_name === Directives.QAPLA) {
       new DirQapla(context).execute(directive, async () => {
+        let next_dir = await this.nextDirective(this.directives);
+        this.process(next_dir);
+      })
+    }
+    else if (directive_name === Directives.MAKE) {
+      new DirMake(context).execute(directive, async () => {
         let next_dir = await this.nextDirective(this.directives);
         this.process(next_dir);
       })