+++ tybotRoute/tiledeskChatbotPlugs/directives/DirSetAttribute.js
@@ -76,7 +76,6 @@ class DirSetAttribute {
             throw new Error('context object is mandatory.');
         }
         this.context = context;
-        this.tdcache = context.tdcache;
         this.log = context.log;
     }
 
@@ -108,71 +107,24 @@ class DirSetAttribute {
             callback();
             return;
         }
+
         
         if (action.operation.operators !== undefined && action.operation.operators.length !== action.operation.operands.length - 1) {
             if (this.log) {console.error("(DirSetAttribute) Invalid action: operators.length !== operands.length - 1")};
             callback();
             return;
         }
-        console.log("filling in setattribute...");
-        await this.fillValues(action.operation.operands);
+        
+
         const expression = TiledeskExpression.JSONOperationToExpression(action.operation.operators, action.operation.operands);
         const attributes = await TiledeskChatbot.allParametersStatic(this.context.tdcache, this.context.requestId);
         attributes.TiledeskMath = TiledeskMath;
         attributes.TiledeskString = TiledeskString;
-        
-        const result = new TiledeskExpression().evaluateJavascriptExpression(expression, attributes);
-        console.log("filling in setattribute, result:", result);
-        await TiledeskChatbot.addParameterStatic(this.context.tdcache, this.context.requestId, action.destination, result);
-        callback();
-    }
 
-    async fillValues(operands) {
-        // operation: {
-        //     operators: ["addAsNumber", "subtractAsNumber", "divideAsNumber", "multiplyAsNumber"],
-        //     operands: [
-        //         {
-        //             value: "previous",
-        //             isVariable: true
-        //         },
-        //         {
-        //             value: "temp",
-        //             isVariable: true,
-        //             function: "floorAsNumber"
-        //         },
-        //         {
-        //             value: "real",
-        //             isVariable: true,
-        //             function: "absAsNumber"
-        //         },
-        //         {
-        //             value: "input",
-        //             isVariable: true
-        //         },
-        //         {
-        //             value: "2",
-        //             isVariable: false
-        //         }
-        //     ]
-        try {
-            if (this.tdcache) {
-                console.log("tdcache in setattribute...", this.tdcache);
-                const requestAttributes = 
-                    await TiledeskChatbot.allParametersStatic(this.tdcache, this.context.requestId);
-                console.log("requestAttributes in setattribute...", requestAttributes);
-                const filler = new Filler();
-                operands.forEach(operand => {
-                    if (!operand.isVariable) {
-                        console.log("setattribute, liquid operand:", operand);
-                        operand.value = filler.fill(operand.value, requestAttributes);
-                        console.log("setattribute, final operand:", operand);
-                    }
-                });
-            }
-        }
-        catch(error) {
-            console.error("Error while filling operands:", error);
-        }
+        const result = new TiledeskExpression().evaluateJavascriptExpression(expression, attributes);
+        await TiledeskChatbot.addParameterStatic(this.context.tdcache, this.context.requestId, action.destination, result);
+
+        callback();
     }
 }